Make MicroContext abstract
MicroContext serves as the primary API between the interpreter and
kernels. For interpreter-less execution, we need to be able to
substitute a different MicroContext implementation, as there will be no
allocator or model.

This PR refactors MicroContext to mostly become an abstract base class,
with existing functionality moved to MicroInterpreterContext, which is a
subclass. A few functions remained, as they were just wrappers around
other MicroContext functions. This will allow us to substitute in an
inference-only version of MicroContext for codegen, which could stub out
all of the allocation methods and re-implement the rest of the
functionality. A downside to this approach is that we'll now have a
dynamic dispatch on each MicroContext access.

FakeMicroContext was also updated to use the MicroContext base class.

BUG=b/295174086